Simulation Object Restructure
